RegistryCache object moved into RegistryCacheConfigSpec
